build ok

+++ vs-linux/src/vport.h
@@ -204,7 +204,7 @@ void ovs_vport_deferred_free(struct vport *vport);
 static inline void *vport_priv(const struct vport *vport)
 {
 	//return (u8 *)(uintptr_t)vport  /+ *ALIGN(sizeof(struct vport), VPORT_ALIGN) TODO just for compile OK , ALIGN issue*/;
-	return (u8 *)(uintptr_t)vport + ((((sizeof(struct vport))) + ((typeof(sizeof(struct vport))) (8) - 1)) & ~((typeof(sizeof(struct vport))) (8) - 1));
+	return (u8 *)vport + ((((sizeof(struct vport))) + ((typeof(sizeof(struct vport))) (8) - 1)) & ~((typeof(sizeof(struct vport))) (8) - 1));
 
 }
